refactor: svg 파일 충돌로 인해 svg 파일 추가, 수정
