+++ src/components/pages/main/group-study-list-page/group-study-thumbnail/GroupStudyRecommendThumbnail.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import * as style from "@/components/pages/main/group-study-page/group-study-thumbnail/GroupStudyRecommendThumbnail.style";
+import * as style from "@/components/pages/main/group-study-list-page/group-study-thumbnail/GroupStudyRecommendThumbnail.style";
 import Image from "next/image";
 import { groupStudyRecommendList } from "@/apis/group-study/group-study-list/groupStudyList";
 
@@ -62,9 +62,9 @@ export default function GroupStudyRecommendThumbnail() {
               </style.RecommendHashTag>
               <style.RecommendViews>
                 <style.RecommendViewsImage
-                  src="/assets/img/Views.svg"
+                  src="/assets/img/GroupStudy_Views.svg"
                   alt="Views"
-                  loader={() => "/assets/img/Views.svg"}
+                  loader={() => "/assets/img/GroupStudy_Views.svg"}
                   width={19}
                   height={17}
                 />
